add some logging

+++ cads-edge/src/upload.cpp
@@ -6,11 +6,14 @@
 
 #include <date/date.h>
 #include <date/tz.h>
+#include <spdlog/spdlog.h>
 
 #include <db.h>
 #include <constants.h>
 #include <coms.h>
 #include <msg.h>
+
+
 
 namespace cads
 {
@@ -27,14 +30,19 @@ bool alarm() {
 
 int resume_scan(std::string scan_name)
 {
+  spdlog::get("cads")->info("Posting a scan. {}", scan_name);
   auto [t,err] = post_scan(scan_name);
-  delete_scan_state(scan_name);
+  if(!err) {
+    delete_scan_state(scan_name);
+    std::filesystem::remove(scan_name);
+    spdlog::get("cads")->info("Removing a posted scan. {}", scan_name);
+  }
   return 0;
 }
 
 void upload_scan_thread(moodycamel::BlockingReaderWriterQueue<msg> &fifo) 
 {
-  std::future<std::invoke_result_t<decltype(resume_scan), std::string >> fut;
+  std::future<std::invoke_result_t<decltype(resume_scan), std::string>> fut;
   std::list<decltype(fut)> running_uploads;
 
   bool loop = true;
@@ -47,13 +55,14 @@ void upload_scan_thread(moodycamel::BlockingReaderWriterQueue<msg> &fifo)
     // keep last scan
     while(scans.size() > 1)
     {
-      auto [scan_begin,path,uploaded] = scans.front();
+      auto [scan_begin,path,uploaded,status] = scans.front();
       scans.pop_front();
       
-      if(uploaded == 0) {
+      if(uploaded == 0 && status == 0) {
        
         delete_scan_state(path);
         std::filesystem::remove(path);
+        spdlog::get("cads")->info("Removing a scan. {}", path);
       }else {
         running_uploads.push_back(std::async(resume_scan, path));
       }
@@ -63,8 +72,9 @@ void upload_scan_thread(moodycamel::BlockingReaderWriterQueue<msg> &fifo)
     std::erase_if(running_uploads,[](const decltype(fut) &f ){return f.wait_for(std::chrono::seconds(0)) == std::future_status::ready;});
 
     
-    if(scans.size() > 0 && true) {
-      auto [scan_begin,path,uploaded] = scans.front();
+    if(scans.size() > 0 && alarm() && running_uploads.size() == 0) {
+      auto [scan_begin,path,uploaded,status] = scans.front();
+      store_scan_status(1,path);
       running_uploads.push_back(std::async(resume_scan, path));
     }
 
@@ -76,6 +86,7 @@ void upload_scan_thread(moodycamel::BlockingReaderWriterQueue<msg> &fifo)
           loop = false;
           break;
         default:
+          spdlog::get("cads")->info("Recieved a scan");
           break;
       }
